requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.9
google-genai==1.13.0
google-api-python-client==2.169.0
google-auth-oauthlib==1.2.2
icalendar==6.1.1
pydantic-settings==2.9.1
